Refactor...! +update of libraries, etc.

+++ CGS - Orienteering Device/Core/Inc/xbeeChecksum.h
@@ -8,9 +8,8 @@
 #ifndef INC_XBEECHECKSUM_H_
 #define INC_XBEECHECKSUM_H_
 
-#include "stdbool.h"
 #include "stm32l0xx_hal.h"
 
-uint8_t xbeeChecksum(uint8_t *buffer, uint16_t length);
+uint8_t xbeeChecksum(uint8_t *buffer, int length);
 
 #endif /* INC_XBEECHECKSUM_H_ */
